+++ mm2src/coins/tendermint/iris/htlc.rs
@@ -2,7 +2,25 @@
 //
 // This module includes HTLC creating & claiming representation structstures
 // and their trait implementations.
+//
+// ** Acquiring testnet assets **
+//
+// Since there is no sdk exists for Rust on Iris Network, we should
+// either implement some of the Iris Network funcionality on Rust or
+// simply use their unit tests.
+//
+// Because we had limited time for the HTLC implementation, for now
+// we can use their unit tests in order to acquire IBC assets.
+// For that, clone https://github.com/ozkanonur/irishub-sdk-js repository and check
+// dummy.test.ts file(change the asset, amount, target address if needed)
+// and then run the following commands:
+// - yarn
+// - npm run test
+//
+// If the sender address doesn't have enough nyan tokens to complete unit tests,
+// check this page https://www.irisnet.org/docs/get-started/testnet.html#faucet
 
+use super::htlc_proto::{ClaimHtlcProtoRep, CreateHtlcProtoRep};
 use cosmrs::{tx::{Fee, Msg, MsgProto},
              AccountId, Coin, ErrorReport};
 use std::convert::TryFrom;
@@ -20,29 +38,6 @@ pub(crate) struct IrisHtlc {
 
     /// Message payload to be sent
     pub(crate) msg_payload: cosmrs::Any,
-}
-
-/// Proto representation of create HTLC message
-#[derive(prost::Message)]
-pub(crate) struct CreateHtlcProtoRep {
-    #[prost(string, tag = "1")]
-    pub(crate) sender: prost::alloc::string::String,
-    #[prost(string, tag = "2")]
-    pub(crate) to: prost::alloc::string::String,
-    #[prost(string, tag = "3")]
-    pub(crate) receiver_on_other_chain: prost::alloc::string::String,
-    #[prost(string, tag = "4")]
-    pub(crate) sender_on_other_chain: prost::alloc::string::String,
-    #[prost(message, repeated, tag = "5")]
-    pub(crate) amount: prost::alloc::vec::Vec<cosmrs::proto::cosmos::base::v1beta1::Coin>,
-    #[prost(string, tag = "6")]
-    pub(crate) hash_lock: prost::alloc::string::String,
-    #[prost(uint64, tag = "7")]
-    pub(crate) timestamp: u64,
-    #[prost(uint64, tag = "8")]
-    pub(crate) time_lock: u64,
-    #[prost(bool, tag = "9")]
-    pub(crate) transfer: bool,
 }
 
 #[derive(Clone)]
@@ -125,17 +120,6 @@ impl From<&MsgCreateHtlc> for CreateHtlcProtoRep {
 
 impl MsgProto for CreateHtlcProtoRep {
     const TYPE_URL: &'static str = CREATE_HTLC_TYPE_URL;
-}
-
-/// Proto representation of claim HTLC message
-#[derive(prost::Message)]
-pub(crate) struct ClaimHtlcProtoRep {
-    #[prost(string, tag = "1")]
-    pub(crate) sender: prost::alloc::string::String,
-    #[prost(string, tag = "2")]
-    pub(crate) id: prost::alloc::string::String,
-    #[prost(string, tag = "3")]
-    pub(crate) secret: prost::alloc::string::String,
 }
 
 #[derive(Clone)]
