+++ mm2src/coins/tendermint/mod.rs
@@ -1,7 +1,9 @@
+// Module implementing Tendermint (Cosmos) integration
+// Useful resources
+// https://docs.cosmos.network/
+
 #[path = "iris/htlc.rs"] mod htlc;
-/// Module implementing Tendermint (Cosmos) integration
-/// Useful resources
-/// https://docs.cosmos.network/
+#[path = "iris/htlc_proto.rs"] mod htlc_proto;
 mod tendermint_coin;
 #[cfg(not(target_arch = "wasm32"))] mod tendermint_native_rpc;
 #[cfg(target_arch = "wasm32")] mod tendermint_wasm_rpc;
